Permito cargar toneladas con decimales en los viajes de cereales
El campo toneladas pasa de entero a Decimal con dos decimales (migracion 0043), siguiendo la misma convencion que el resto de las cantidades comerciales del sistema para evitar el ruido de precision del punto flotante. La validacion acepta coma o punto como separador, exige un valor positivo, rechaza mas de dos decimales y normaliza el resultado, de modo que 5,5 se guarda como 5.50. Cubre tanto la creacion como la edicion, que comparten el mismo validador.

En los formularios de nuevo acopio y de edicion, el campo pasa de type number a type text con inputmode decimal y un pattern que admite hasta dos decimales, con el mensaje de error actualizado.

+++ main/models.py
@@ -431,7 +431,10 @@ class ViajeCereal(models.Model):
     # El CTG es un codigo de hasta 15 digitos que puede tener ceros a la izquierda, por eso
     # lo guardo como texto: un IntegerField perderia esos ceros (00123456 -> 123456)
     codigo_trazabilidad_granos = models.CharField(max_length=15)
-    toneladas = models.PositiveIntegerField(default=0)
+    # Toneladas transportadas. Uso Decimal (no Integer/Float) para admitir hasta dos
+    # decimales sin el ruido de precision del punto flotante, igual que en el resto de
+    # las cantidades comerciales del sistema (DetalleOperacion, Cotizaciones).
+    toneladas = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     precio_tonelada = models.PositiveIntegerField(default=0)
     porcentaje_chofer = models.PositiveIntegerField(default=0)
     activo = models.BooleanField(default=True)
